backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
rembg>=2.0.68
pillow==10.4.0
prometheus-client==0.21.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
aiohttp==3.10.5
requests==2.32.3
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.2
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
